Remove SimpleEmitter (close #341)
